# src/main/java/com/sj/yets/api/BasicInfoApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sj.yets.api;

import com.sj.yets.api.model.BasicInfoDTO;
import java.util.List;
import com.sj.yets.api.model.YetsResultDTO;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Api(value = "BasicInfo", description = "the BasicInfo API")
public interface BasicInfoApi {

    /**
     * GET /api/v1/basicInfo/calculate/{name}/{jumin}
     *
     * @param name name (required)
     * @param jumin jumin (required)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "calculateYets", notes = "", response = YetsResultDTO.class, responseContainer = "List", tags={ "basicInfo", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = YetsResultDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/basicInfo/calculate/{name}/{jumin}",
        produces = { "application/json" }
    )
    ResponseEntity<List<YetsResultDTO>> calculateYets(@ApiParam(value = "name", required = true) @PathVariable("name") String name,@ApiParam(value = "jumin", required = true) @PathVariable("jumin") String jumin);


    /**
     * GET /api/v1/basicInfo
     *
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "getBasicInfo", notes = "", response = BasicInfoDTO.class, responseContainer = "List", tags={ "basicInfo", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = BasicInfoDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/basicInfo",
        produces = { "application/json" }
    )
    ResponseEntity<List<BasicInfoDTO>> getBasicInfo();


    /**
     * GET /api/v1/basicInfo/{name}/{jumin}
     *
     * @param name name (required)
     * @param jumin jumin (required)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "getBasicInfoByParam", notes = "", response = BasicInfoDTO.class, responseContainer = "List", tags={ "basicInfo", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = BasicInfoDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/basicInfo/{name}/{jumin}",
        produces = { "application/json" }
    )
    ResponseEntity<List<BasicInfoDTO>> getBasicInfoByParam(@ApiParam(value = "name", required = true) @PathVariable("name") String name,@ApiParam(value = "jumin", required = true) @PathVariable("jumin") String jumin);


    /**
     * POST /api/v1/basicInfo
     *
     * @param basicInfoDTO  (optional)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "insertBasicInfo", notes = "", response = BasicInfoDTO.class, responseContainer = "List", tags={ "basicInfo", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = BasicInfoDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/basicInfo",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<List<BasicInfoDTO>> insertBasicInfo(@ApiParam(value = "") @Valid @RequestBody(required = false) List<BasicInfoDTO> basicInfoDTO);

}