# src/main/java/com/sj/yets/api/HumanDeductionApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sj.yets.api;

import com.sj.yets.api.model.HumanDeductionDTO;
import java.util.List;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Api(value = "HumanDeduction", description = "the HumanDeduction API")
public interface HumanDeductionApi {

    /**
     * GET /api/v1/humanDeduction/{jumin}
     *
     * @param jumin jumin (required)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "getHumanDeductionByParam", notes = "", response = HumanDeductionDTO.class, responseContainer = "List", tags={ "humanDeduction", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = HumanDeductionDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/humanDeduction/{jumin}",
        produces = { "application/json" }
    )
    ResponseEntity<List<HumanDeductionDTO>> getHumanDeductionByParam(@ApiParam(value = "jumin", required = true) @PathVariable("jumin") String jumin);


    /**
     * POST /api/v1/humanDeduction/{jumin}
     *
     * @param jumin jumin (required)
     * @param humanDeductionDTO  (optional)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "insertHumanDeduction", notes = "", response = HumanDeductionDTO.class, responseContainer = "List", tags={ "humanDeduction", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = HumanDeductionDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/humanDeduction/{jumin}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<List<HumanDeductionDTO>> insertHumanDeduction(@ApiParam(value = "jumin", required = true) @PathVariable("jumin") String jumin,@ApiParam(value = "") @Valid @RequestBody(required = false) List<HumanDeductionDTO> humanDeductionDTO);

}