# src/main/java/com/sj/yets/api/IncomeDeductionApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sj.yets.api;

import com.sj.yets.api.model.IncomeDeductionDTO;
import java.util.List;
import com.sj.yets.api.model.YetsResultDTO;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Api(value = "IncomeDeduction", description = "the IncomeDeduction API")
public interface IncomeDeductionApi {

    /**
     * GET /api/v1/incomeDeduction/{jumin}
     *
     * @param jumin jumin (required)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "getIncomeDeduction", notes = "", response = IncomeDeductionDTO.class, responseContainer = "List", tags={ "incomeDeduction", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = IncomeDeductionDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/incomeDeduction/{jumin}",
        produces = { "application/json" }
    )
    ResponseEntity<List<IncomeDeductionDTO>> getIncomeDeduction(@ApiParam(value = "jumin", required = true) @PathVariable("jumin") String jumin);


    /**
     * GET /api/v1/incomeDeductionAmount/{jumin}
     *
     * @param jumin jumin (required)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "getIncomeDeductionAmount", notes = "", response = YetsResultDTO.class, responseContainer = "List", tags={ "incomeDeduction", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = YetsResultDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/incomeDeductionAmount/{jumin}",
        produces = { "application/json" }
    )
    ResponseEntity<List<YetsResultDTO>> getIncomeDeductionAmount(@ApiParam(value = "jumin", required = true) @PathVariable("jumin") String jumin);


    /**
     * GET /api/v1/incomeDeductionLimit/{jumin}
     *
     * @param jumin jumin (required)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "getIncomeDeductionLimit", notes = "", response = YetsResultDTO.class, responseContainer = "List", tags={ "incomeDeduction", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = YetsResultDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/incomeDeductionLimit/{jumin}",
        produces = { "application/json" }
    )
    ResponseEntity<List<YetsResultDTO>> getIncomeDeductionLimit(@ApiParam(value = "jumin", required = true) @PathVariable("jumin") String jumin);


    /**
     * POST /api/v1/incomeDeduction/{jumin}
     *
     * @param jumin jumin (required)
     * @param incomeDeductionDTO  (optional)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "insertIncomeDeduction", notes = "", response = IncomeDeductionDTO.class, responseContainer = "List", tags={ "incomeDeduction", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = IncomeDeductionDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/incomeDeduction/{jumin}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<List<IncomeDeductionDTO>> insertIncomeDeduction(@ApiParam(value = "jumin", required = true) @PathVariable("jumin") String jumin,@ApiParam(value = "") @Valid @RequestBody(required = false) List<IncomeDeductionDTO> incomeDeductionDTO);

}