# src/main/java/com/sj/yets/api/TaxCreditApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sj.yets.api;

import java.util.List;
import com.sj.yets.api.model.TaxCreditDTO;
import com.sj.yets.api.model.YetsResultDTO;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Api(value = "TaxCredit", description = "the TaxCredit API")
public interface TaxCreditApi {

    /**
     * GET /api/v1/taxCredit/{jumin}
     *
     * @param jumin jumin (required)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "getTaxCredit", notes = "", response = TaxCreditDTO.class, responseContainer = "List", tags={ "taxCredit", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = TaxCreditDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/taxCredit/{jumin}",
        produces = { "application/json" }
    )
    ResponseEntity<List<TaxCreditDTO>> getTaxCredit(@ApiParam(value = "jumin", required = true) @PathVariable("jumin") String jumin);


    /**
     * GET /api/v1/taxCreditAmount/{jumin}
     *
     * @param jumin jumin (required)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "getTaxCreditAmount", notes = "", response = YetsResultDTO.class, responseContainer = "List", tags={ "taxCredit", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = YetsResultDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/taxCreditAmount/{jumin}",
        produces = { "application/json" }
    )
    ResponseEntity<List<YetsResultDTO>> getTaxCreditAmount(@ApiParam(value = "jumin", required = true) @PathVariable("jumin") String jumin);


    /**
     * GET /api/v1/taxCreditLimit/{jumin}
     *
     * @param jumin jumin (required)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "getTaxCreditLimit", notes = "", response = YetsResultDTO.class, responseContainer = "List", tags={ "taxCredit", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = YetsResultDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/taxCreditLimit/{jumin}",
        produces = { "application/json" }
    )
    ResponseEntity<List<YetsResultDTO>> getTaxCreditLimit(@ApiParam(value = "jumin", required = true) @PathVariable("jumin") String jumin);


    /**
     * POST /api/v1/taxCredit/{jumin}
     *
     * @param jumin jumin (required)
     * @param taxCreditDTO  (optional)
     * @return Success (status code 200)
     */
    @ApiOperation(value = "", nickname = "insertTaxCredit", notes = "", response = TaxCreditDTO.class, responseContainer = "List", tags={ "taxCredit", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success", response = TaxCreditDTO.class, responseContainer = "List") })
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/taxCredit/{jumin}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<List<TaxCreditDTO>> insertTaxCredit(@ApiParam(value = "jumin", required = true) @PathVariable("jumin") String jumin,@ApiParam(value = "") @Valid @RequestBody(required = false) List<TaxCreditDTO> taxCreditDTO);

}